Adding german language version of the word clock.
Based on wordclock-english/wordclock-english.ino a version
wordclock-german/wordclock-german.ino was created for the german version
of Emily Hammes Word Clock.

This version of the code already contains the code for the two idioms
that exist in german: "viertel vor eins" vs. "dreiviertel eins". When
powerd on the button of the clock controls which idiom is used.

Some changes where made to the naming of the variables and functions
in order to better distinguishe between global and local variables and
those provided as a function parameter. In addition comments are added
to the code to explain what's going on.

The original (english) version is untouched.